pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.3",
    "openai>=1.107.1",
    "paho-mqtt>=2.1.0",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "streamlit>=1.49.1",
    "twilio>=9.8.0",
]
